Simplify getArgs()
getArgs() no longer returns a package name, and a type of operation, they are now determined in main by checking if the value of install or uninstall are different from "". Although it's still not the best way to handle that, it allows for easier handling of arguments outside getArgs(), this function will now parse and check the validity of the arguments.