Add features to install and uninstall packages as well as managing the repo
Added database.go, which contains functions and methods to manipulate the database. The comments are self-explanatory.

Added add.go which handles the adding of a package in the database, comments are self-explanatory.

Added manage.go to manage packages that are already in the database, it includes the modification of every aspect of a package record in the database.

Added util.go which contains pre-defined colors for the installation/uninstallation steps output. There's the CheckRoot() function which checks if the user is root and prints an error if the user is not root. There's also the Display() functions which prints to specified output.

manage.go, install.go and uninstall.go are part of the same package. I don't like the way it is currently and this is subject to change.

Added manage.go to prep for installation and uninstallation including a function to ask verify and ask before installing/uninstalling which can be bypasses with --yes/-y.

Added install.go which has a function to install packages using functions in installation scripts and their Makefile.

Added uninstall.go which has a function to uninstall packages using functions in installation scripts and their Makefile.

Added more flags to root.go, comments for these are self-explanatory.

Added more flags to show.go, to show general information about packages which includes their licenses, name, installed or not, installed version (if installed), description, repo's version. There's also --all/-a to show every packages general information.

Removed testing because testing is doubting.