Separate the subcommands logic into separate packages for each
The logic used in subcommands were separated into their own packages, for each, there's a Decide() function that determines what to do (checking the values of string and booleans), connects to the database, then perform the actions of the subcommands.

The logic is copy-pasted from the original one but separated into functions for the main actions that are called by the Decide() function. Nothing really changed apart from adding the removal of trailing slashes for the buildFilesDir.