Refactor pkg to ressemble the other packages
The package pkg has been refactored to ressemble the other packages with a Decide() function that calls function that make the operations based on the given options.

The installation process does not depend on a bash script anymore, juste on a makefile, the downloading, verification, and extracting are done internally.

Also added an experimental dependency resolution function that can be used optionnaly with --resolve.

The comments are self-explanatory.